refactor(priming-group): move pipeline and priming group methods to ExperimentalApi
